refactor: Video 중복타입 정리
videoApi.ts: Video 중복타입 정리
videoSlice.ts: Video 중복타입 정리
Video.ts: Video type
tsconfig.json: include ["src"] -> include ["src", "src/**/*"]
